Gasoline-values should be non-negative integers

+++ funsearch/gasoline/iterative_rounding.py
@@ -26,7 +26,7 @@ class IterativeAlgo(abc.ABC) :
     def run(self, instance : ins.Instance) -> tuple[Result, float] :
         pass
 
-    def approximation_ratio(self, xs: List[float], ys: List[float]) -> float :
+    def approximation_ratio(self, xs: List[int], ys: List[int]) -> float :
           n = len(xs)
           if len(ys) < n - 1:
               print(f"<*> len(ys) < n-1")
